+++ src/main/java/me/dartanman/duels/commands/subcommands/duel/LeaveDuelsSubCmd.java
@@ -1,6 +1,7 @@
-package me.dartanman.duels.commands.subcommands;
+package me.dartanman.duels.commands.subcommands.duel;
 
 import me.dartanman.duels.Duels;
+import me.dartanman.duels.commands.subcommands.DuelsSubCommand;
 import me.dartanman.duels.game.GameState;
 import me.dartanman.duels.game.arenas.Arena;
 import me.dartanman.duels.utils.PlayerRestoration;
